Re-arrange component composition for SelectInstrument
- Separate selections from SelectInstrument
- Move the StartButton from Room to be a sub-component of SelectInstrument
- Ensure handlers are being passed down as props, to correct place
- Ensure props are being passed down
